Perda estoque
This project aims to work on data provided by SEARA ALIMENTOS LTDA, so we will deal with real market data, protecting privacy and sensitive content previously informed by the company. Throughout the project, we will seek to efficiently present results for the business problem, and show how to apply this solution in order to obtain results in a timely manner in the company. At this stage it is interesting to explain a bit of context in which your project is inserted. You can talk about the market situation, trends that are important to observe and the main challenges that exist. If they are doing work for a specific company, they can provide a brief overview of the company as well.